# Restore optimizations for `NDBuffer.all_equal`